Add support of custom pcg painter actor class
User can now select the class of pcg painter actor to paint.
This allows the user to create subclasses ( of PCGPainterActor ) which could have variables exposed to be read by a PCG graph using GetActorProperty node, the graph selection will be hidden as the graph would be embedded in the BP class directly.
All variables set in the category "Painter Parameters" will be shown if "show instance parameters" is checked on.
If the class is set to PCGPainterActor ( by default ), then the behavior doesn't change, the user need to select the graph processor to paint.